feat: move helper and update genesis account
