+++ node/Cargo.toml
@@ -81,6 +81,7 @@ stability-runtime = { path = "./../runtime", features = ["std"] }
 # Stability
 stbl-cli-authorship = { workspace = true }
 stbl-primitives-fee-compatible-api = { workspace = true, features = ["std"] }
+stbl-tools = { workspace = true }
 
 [build-dependencies]
 substrate-build-script-utils = { workspace = true }
