+++ primitives/tools/Cargo.toml
@@ -12,3 +12,4 @@ sp-runtime = { workspace = true }
 sp-std = { workspace = true }
 sp-io = { workspace = true }
 ethereum = { workspace = true }
+sha3 = { version = "0.10.6", default-features = false }
